# api/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "LoanLens API",
    "description": "Advanced Loan Default Prediction API with explainable AI, model management, A/B testing, drift detection, and automated retraining.",
    "version": "2.0.0",
    "contact": {
      "name": "LoanLens Development Team",
      "email": "support@loanlens.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    },
    {
      "url": "https://api.loanlens.ai/v2",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/predict": {
      "post": {
        "summary": "Make Prediction",
        "description": "Makes a single loan default prediction with probability and SHAP explanations",
        "operationId": "predict",
        "tags": ["Predictions"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prediction result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing features or no models loaded"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/ab-tests": {
      "post": {
        "summary": "Create A/B Test",
        "description": "Creates a new A/B test comparing two model versions",
        "operationId": "createABTest",
        "tags": ["A/B Testing"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ABTestCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Test created successfully"
          }
        }
      },
      "get": {
        "summary": "List A/B Tests",
        "description": "Returns all A/B tests",
        "operationId": "listABTests",
        "tags": ["A/B Testing"],
        "responses": {
          "200": {
            "description": "List of tests"
          }
        }
      }
    },
    "/ab-tests/{test_id}": {
      "get": {
        "summary": "Get A/B Test Status",
        "description": "Returns detailed status of an A/B test",
        "operationId": "getABTestStatus",
        "tags": ["A/B Testing"],
        "parameters": [
          {
            "name": "test_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Test status"
          },
          "404": {
            "description": "Test not found"
          }
        }
      }
    },
    "/ab-tests/{test_id}/start": {
      "post": {
        "summary": "Start A/B Test",
        "description": "Starts a scheduled A/B test",
        "operationId": "startABTest",
        "tags": ["A/B Testing"],
        "parameters": [
          {
            "name": "test_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Test started"
          }
        }
      }
    },
    "/ab-tests/{test_id}/analyze": {
      "get": {
        "summary": "Analyze A/B Test",
        "description": "Statistical analysis with p-values and effect sizes",
        "operationId": "analyzeABTest",
        "tags": ["A/B Testing"],
        "parameters": [
          {
            "name": "test_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          },
          {
            "name": "metric",
            "in": "query",
            "schema": {"type": "string", "default": "prediction"}
          }
        ],
        "responses": {
          "200": {
            "description": "Statistical analysis results"
          }
        }
      }
    },
    "/drift/detect": {
      "post": {
        "summary": "Detect Data Drift",
        "description": "Detects data drift in uploaded dataset using KS test, PSI, or Wasserstein distance",
        "operationId": "detectDrift",
        "tags": ["Drift Detection"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {"type": "string", "format": "binary"},
                  "method": {"type": "string", "enum": ["ks_test", "psi", "wasserstein"]}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Drift detection results"
          }
        }
      }
    },
...
    "/models/promote": {
      "post": {
        "summary": "Promote Model",
        "description": "Promotes a model version to production",
        "operationId": "promoteModel",
        "tags": ["Model Management"],
        "parameters": [
          {
            "name": "model_name",
            "in": "query",
            "required": true,
            "schema": {"type": "string"}
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Model promoted"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PredictionRequest": {
        "type": "object",
        "required": ["features"],
        "properties": {
          "features": {
            "type": "object",
            "description": "Feature values for prediction"
          },
          "model_name": {"type": "string"},
          "threshold": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "default": 0.5
          }
        }
      },
      "PredictionResponse": {
        "type": "object",
        "properties": {
          "prediction": {"type": "integer"},
          "probability": {"type": "number"},
          "model_name": {"type": "string"},
          "threshold": {"type": "number"},
          "timestamp": {"type": "string"},
          "explanation": {"type": "object"},
          "latency_ms": {"type": "number"}
        }
      },
      "ABTestCreateRequest": {
        "type": "object",
        "required": ["name", "model_a_name", "model_a_version", "model_b_name", "model_b_version"],
        "properties": {
          "name": {"type": "string"},
          "model_a_name": {"type": "string"},
          "model_a_version": {"type": "string"},
          "model_b_name": {"type": "string"},
          "model_b_version": {"type": "string"},
          "traffic_split": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "default": 0.5
          }
        }
      }
    }
  }
}